% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_plotter.R
\name{chi_square_plotter}
\alias{chi_square_plotter}
\title{Chi Square Plotter
This function quickly plots a summarised plot for outlier detection with numeric variables}
\usage{
chi_square_plotter(x)
}
\arguments{
\item{x}{a dataframe with named rows}
}
\description{
Chi Square Plotter
This function quickly plots a summarised plot for outlier detection with numeric variables
}
\keyword{chi_square}
\keyword{multivariate}
