% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_chi_square.R
\name{compare_chi_square}
\alias{compare_chi_square}
\title{compare_chi_square}
\usage{
compare_chi_square(df, group_1, response)
}
\arguments{
\item{df}{a data frame}

\item{group_1}{first grouping variable}

\item{response}{treament that you are comparing (binary)}
}
\description{
compare_chi_square
}
