% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_box_legend.R
\name{ggplot_box_legend}
\alias{ggplot_box_legend}
\title{ggplot_box_legend}
\usage{
ggplot_box_legend(family = "serif")
}
\arguments{
\item{family}{default = "serif"}
}
\description{
Boxplot explanation from USGS https://owi.usgs.gov/blog/boxplots/
}
