% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_depedencies.R
\name{install_dep}
\alias{install_dep}
\title{Install Dependent Packages}
\usage{
install_dep(pkg.dir = ".", dependencies = TRUE,
  repos = getOption("repos")[1], lib = .libPaths()[1])
}
\arguments{
\item{pkg.dir}{refers to the package directory that contains the
\code{Description} file}

\item{dependencies}{defines which dependencies of the dependent packages are
to be installed}

\item{repos}{is the (CRAN) repository used to install dependencies}

\item{lib}{is the library to which packages are installed}
}
\description{
Install Dependent Packages
}
