% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_survey_key.R
\name{make_survey_key}
\alias{make_survey_key}
\title{make survey key}
\usage{
make_survey_key(survey)
}
\arguments{
\item{survey}{a list item that has been created from the qualtrics API}
}
\description{
This is a function that can be used to extract information from a survey item created by the qualtrics API
}
