% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wellbeing_chart.R
\name{make_wellbeing_charts}
\alias{make_wellbeing_charts}
\title{MAke Wellbeing Likert Chart}
\usage{
make_wellbeing_charts(data, mygroup, likert_scale)
}
\arguments{
\item{mygroup}{the question name you wish to group}

\item{likert_scale}{the Likert Scale to Use ("agreement", "frequency", "satisfaction")}
}
\details{
The purpose of this function to faciliate making charts quickly
it utilizes the stock likert scales that are used in the wellbeing survey
}
