% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_effect_size_2.R
\name{my_es_2}
\alias{my_es_2}
\title{My Effect Size Calculation Two}
\usage{
my_es_2(df)
}
\arguments{
\item{df}{with a column titled key, my, var and freq}
}
\description{
A simple tool to calculate pairwise effect sizes
  It takes the dataframe that has already been gathered (using tidyr)
  and summarised into variables that include mu(mean), var(variance) and
  freq(count). It then applies Cohen's D equation to calculate effect size
  and output both the scaled effect and the descriptor.

  It will apply this to data that is given in the "key", "value" format.
}
