% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_scale.R
\name{range_scale}
\alias{range_scale}
\title{A function to scale values based on ranges rather than typical R scale function}
\usage{
range_scale(df)
}
\arguments{
\item{df}{which contains the numeric values to be scaled via range}
}
\value{
df with the scaled values
}
\description{
A function to scale values based on ranges rather than typical R scale function
}
