% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_colour_wake}
\alias{scale_colour_wake}
\alias{scale_color_wake}
\alias{scale_fill_wake}
\title{Discrete color & fill scales based on the ipsum palette}
\usage{
scale_colour_wake(...)

scale_color_wake(...)

scale_fill_wake(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
See \url{wake_pal}.
}
