% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellness_pal.R
\name{scale_colour_wellness}
\alias{scale_colour_wellness}
\alias{scale_color_wellness}
\alias{scale_fill_wellness}
\title{Discrete color & fill scales based on the wellness palette}
\usage{
scale_colour_wellness(...)

scale_color_wellness(...)

scale_fill_wellness(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
See \url{wellness_pal}.
}
