% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_test.R
\name{which_test}
\alias{which_test}
\title{What Frequentist Test?}
\usage{
which_test(data, variable, alpha = 0.05)
}
\arguments{
\item{data}{your dataset to use}

\item{variable}{the variable in which you are interested}
}
\description{
This is a function that tests for normality and then proposes
what tool to use for significance testing (Null Hypothesis Testing)
}
