% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_slurm.R
\name{write_slurm}
\alias{write_slurm}
\title{write_slurm}
\usage{
write_slurm(r_file_name, job_name, cores = 4, mem = 16,
  my_directory = "/deac/generalGrp/usershare/dewittme")
}
\arguments{
\item{r_file_name}{the file R file that you want to use}

\item{job_name}{what do you want to call the job name (recommend using R name without R)}

\item{cores}{how many cores per node to request (default is 4)}

\item{mem}{how many GB of memory to request (default = 16)}

\item{my_directory}{the directory where to store and access your files (default is dewittme)}
}
\value{
a slurm script that can be submitted to the cluster
}
\description{
write_slurm
}
