% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_var.R
\name{wtd_var}
\alias{wtd_var}
\title{weighted variance from Hmisc}
\usage{
wtd_var(x, weights = NULL, normwt = FALSE, na.rm = TRUE,
  method = c("unbiased", "ML"))
}
\arguments{
\item{x}{the vector to average}

\item{weights}{the weights}
}
\description{
weighted variance from Hmisc
}
