% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_plotter.R
\name{chi_square_plotter}
\alias{chi_square_plotter}
\title{Chi Square Plotter}
\usage{
chi_square_plotter(input_df)
}
\arguments{
\item{input_df}{a dataframe with named rows}
}
\description{
Chi Square Plotter
}
